Fix TextScreen URE bug
Oh boy, this one is !!!FUN!!!

Unity sets the field to some mysterious "Null but not really null" value, overrides == and then calls it a day. I come along and check against "is null" and the check fails because it's not ACTUALLY null, just a facsimile of it.

NONE OF THIS is transparent to the programmer. Great!

C# : "Let's take the concept of side effects and make it into a language"